requirements.txt
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.4
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
TestIt==0.1.0
testit-api-client==5.0.0
testit-cli==2.1.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==2.0.2
datafaker==0.7.6
pip~=24.3.1
requests~=2.32.3
beautifulsoup4~=4.13.3
configparser~=7.1.0
cryptography~=44.0.0
lxml~=5.3.1